refactor admin controller directory
