+++ presentation/src/main/java/team/themoment/officialgsm/admin/auth/controller/AuthController.java
@@ -1,4 +1,4 @@
-package team.themoment.officialgsm.client.auth.controller;
+package team.themoment.officialgsm.admin.auth.controller;
 
 
 import org.springframework.web.bind.annotation.RestController;
