+++ presentation/src/main/java/team/themoment/officialgsm/admin/auth/controller/manager/CookieManager.java
@@ -1,4 +1,4 @@
-package team.themoment.officialgsm.admin.auth.manager;
+package team.themoment.officialgsm.admin.auth.controller.manager;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
